# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.Company:
    properties:
      created_at:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  entity.Item:
    properties:
      category_name:
        type: string
      code:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      item_id:
        type: string
      name:
        type: string
      price:
        type: number
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  v1.CreateItemRequest:
    properties:
      category_name:
        type: string
      code:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      type:
        enum:
        - PRODUCT
        - SERVICE
        type: string
    required:
    - code
    - name
    - price
    - type
    type: object
  v1.Response:
    properties:
      code:
        type: string
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  v1.UpdateItemRequest:
    properties:
      category_name:
        type: string
      code:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      status:
        enum:
        - ACTIVE
        - INACTIVE
        type: string
      type:
        enum:
        - PRODUCT
        - SERVICE
        type: string
    required:
    - code
    - name
    - price
    - status
    - type
    type: object
info:
  contact: {}
  description: Backend API for managing products and services.
  title: Product & Service Management API
  version: "1.0"
paths:
  /companies:
    get:
      consumes:
      - application/json
      description: Get a list of all active companies containing uuid and name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Company'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: List companies
      tags:
      - companies
  /companies/{company_id}/items:
    get:
      consumes:
      - application/json
      description: Get a list of items belonging to a company with filters
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      - description: Item Type (PRODUCT, SERVICE)
        in: query
        name: type
        type: string
      - description: Item Status (ACTIVE, INACTIVE, ARCHIVED)
        in: query
        name: status
        type: string
      - description: Search by code or name
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Item'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: List items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new product or service for a company
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      - description: Item creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Item'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Create a new item
      tags:
      - items
  /companies/{company_id}/items/{item_id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific item for a company
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Item'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Get item detail
      tags:
      - items
    patch:
      consumes:
      - application/json
      description: Update a specific item's details. Archived items cannot be updated.
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      - description: Item update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Item'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Update an item
      tags:
      - items
  /companies/{company_id}/items/{item_id}/archive:
    patch:
      consumes:
      - application/json
      description: Archive a specific item. Archived items cannot be archived again.
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Archive an item
      tags:
      - items
schemes:
- http
- https
swagger: "2.0"